did-web3j-sdk/src/main/java/com/reddate/did/protocol/common/BaseCredential.java
// Copyright 2021 Red Date Technology Ltd.  Licensed under MPLv2
// (https://www.mozilla.org/en-US/MPL/2.0/)
package com.reddate.did.protocol.common;

import java.io.Serializable;

public class BaseCredential implements Serializable {

  private String id; // Document No

  private String created; // Creation time

  public String getId() {
    return id;
  }

  public void setId(String id) {
    this.id = id;
  }

  public String getCreated() {
    return created;
  }

  public void setCreated(String created) {
    this.created = created;
  }
}
